+++ packages/flowerbase/src/auth/providers/local-userpass/dtos.ts
@@ -62,3 +62,9 @@ export interface ConfirmUserDto {
     tokenId: string
   }
 }
+
+export interface ConfirmUserSendDto {
+  Body: {
+    email: string
+  }
+}